+++ CHANGELOG.md
@@ -19,6 +19,8 @@ Changes are grouped as follows
 * Support `dev` argument in `resync init` to destroy recreate views and
   data models in the development process.
 * Regenerate clients with `pygen` `v0.32.5`
+* Introduce `PowerOpsContainerModel` for an easy way to see all available
+  containers in `PowerOps`
 
 ## [0.78.0] - 2023-12-22
 ### Changed